backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9  # For PostgreSQL support

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
python-jose[pycryptodome]==3.3.0
pyjwt==2.8.0

# Data Validation
pydantic==2.9.1
pydantic-settings==2.9.1
email-validator==2.1.0.post1

# Payments
stripe==8.0.0

# Hacienda Integration
zeep==4.2.1
xmltodict==0.13.0
cryptography==42.0.5

# Utilities
python-dateutil==2.9.0
pytz==2024.1
requests==2.31.0
typing-extensions==4.12.2

# Templates
Jinja2==3.1.4
